pyproject.toml
[project]
name = "quotes_bot"
version = "0.1.0"
description = "Telegram bot for sending random quotes"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=20.7",
    "python-dotenv>=1.0.0",
    "reportlab>=4.0.0"
]
